Change the structure of ts config
