--- dev-tools/donate-table/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ESNext",
-    "esModuleInterop": true,
-    "module": "commonjs",
-    "strict": true,
-    "sourceMap": false,
-    "skipLibCheck": true
-  },
-  "include": [
-    "index.ts"
-  ]
-}
