--- registry/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "paths": {
-      "@/*": [
-        "src/*"
-      ]
-    }
-  },
-  "include": [
-    "../src/**/*.d.ts",
-    "lib/**/*.ts"
-  ],
-}
