+++ registry/webpack/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../tsconfig.node.json"
+}
