+++ src/core/settings/write.ts
@@ -29,3 +29,4 @@
 //     settingsInternalState.settingsLoaded = true
 //   }
 // }
+export {}
