+++ webpack/webpack-define.d.ts
@@ -1,0 +1,11 @@
+declare global {
+  interface GitInfo {
+    commitHash: string
+    branch: string
+    nearestTag: string
+    versionWithTag: string
+  }
+  const webpackGitInfo: GitInfo
+}
+
+export {}
